Add function to correct gaussian mixture with new measurement
This commit implements the correct function of the gaussian mixture.
The individual mixture components get corrected using Kalman filter
equations. Basic unittests are included.

In the future, the implementation can be further optimized by using
numba and guvectorize in order to enable efficient computation with
broadcasting. The loop can then be reduce to one sequential computation.