Integrate association log-likelihood into correct step
This commit extends the correct function to calculate the
log-likelihood of the current track-measurement association.
Therefore, the detection and misdetection probabilities are
included. In case of a missed detection, only the log-likelihood
is updated, while the mixture itself does not get corrected.

Furthermore, the initialization of calculated parameters is
integrated into the dataclass using the post_init method.
The initialization of log representations directly inside the
Parameters class avoid redundant computations during runtime.