Add adaptive target birth and initial pruning (#15)
This commit adds the adaptive birth function and an initial target pruning based on the existence probability threshold. Required additional parameters are introduced and default birth parameters are initially optimized.

Additionally, the calculation of log-likehood of the hypothesis weights has been rearranged to reduce the required number of computations in different functions.

Furthermore, it refactors the printouts during execution to improve the readability of current targets.