Add merging of close gaussian mixture components
This commit introduces the merge of gaussian mixture
components of a gaussian mixture pdf based on their
(squared) Mahalanobis distance. Components with distance
below a specified value are combined into one component
by a weighed average based on the component weights.

In order to avoid testing dependencies between different
functions, the pruning and merging functions are extracted
into a new reduce_pdf method, which is called by the target
object.