Add estimated number of new-born targets to adaptive birth
The parameter lambda_b corresponds to the estimated number
of new-born targets per measurement update and is added into
the equation of computing the birth probability.
The parameter is set to the mean cardinality of the multi-
Bernoulli RFS of new-born targets, assuming the existence
probability of new-born targets is equal to the probability
of the measurement not being assigned to an existing target.

Additionally, the overall tracker parameter are refined for
improved performance on the 2D point example.